api/requirements.txt
flask
flask_cors
flask_restful
sqlalchemy
mysql-connector
pymysql
requests
Flask-APScheduler
esdk-obs-python
flask_jwt_extended
urllib3
croniter
wrapt
pytz
tzlocal
cryptography
tenacity
omegaconf